requirements.txt
python_json_logger
PyYAML
